gym-sapientino-case-master/pyproject.toml
[tool.poetry]
name = "gym-sapientino-case"
version = "0.2.0"
description = "A specific environment instance created from gym-sapientino"
authors = ["Roberto Cipollone <cipollone.rt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
gym-sapientino = { git = "https://github.com/whitemech/gym-sapientino.git", branch = "data-workaround" }
temprl = { git = "https://github.com/whitemech/temprl.git", branch = "develop" }
logaut = "^0.1.1"

[tool.poetry.dev-dependencies]
ipython = "^7.22.0"
pytest = "^6.2.3"
mypy = "^0.812"
isort = "^5.8.0"
black = "^20.8b1"
jedi = "^0.18.0"
flake8 = "^3.9.0"
tox = "^3.23.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["setuptools", "poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"
